server/requirements.txt
# Core dependencies
flask>=2.0.1
python-socketio>=5.5.0
requests>=2.26.0
Werkzeug>=2.3.7

# Updated for Python 3.13 compatibility
numpy>=1.26.0
opencv-python>=4.8.0
scikit-learn>=1.3.0  # For cosine_similarity

# Face detection and recognition
mediapipe>=0.10.0  # For face detection and alignment
insightface>=0.7.3  # For ArcFace embedding generation
# dlib>=19.24.0; platform_system != "Darwin" or platform_machine != "arm64"
# face-recognition>=1.3.0; platform_system != "Darwin" or platform_machine != "arm64"

# Database
mysql-connector-python>=8.0.27

# Optional ML components - comment these out if installation fails
# ultralytics>=8.0.0  # For YOLOv11x 